reset client init
