+++ src/client/client.ts
@@ -3,8 +3,6 @@ import { KeyDAO } from '../key/keyDAO';
 import * as config from '../config/config';
 import * as sdk from '@irisnet/irishub-sdk';
 
-export var client:sdk.Client;
-
 function createClient(node:any, chainId:any , gas:any, mnemonic:any, algorithm:any): sdk.Client {
   const client = sdk
     .newClient({
@@ -20,7 +18,7 @@ function createClient(node:any, chainId:any , gas:any, mnemonic:any, algorithm:a
   return client;
 }
 
-export function clientInit() {
+export function getClient() {
   let extConfig = vscode.workspace.getConfiguration();
-  client = createClient(extConfig.get("vscode-ext-irita-wasm.rpcAddress"), extConfig.get("vscode-ext-irita-wasm.chainID"), extConfig.get("vscode-ext-irita-wasm.gasLimit"), extConfig.get("vscode-ext-irita-wasm.mnemonic"), extConfig.get("vscode-ext-irita-wasm.algorithm"));
+  return createClient(extConfig.get("vscode-ext-irita-wasm.rpcAddress"), extConfig.get("vscode-ext-irita-wasm.chainID"), extConfig.get("vscode-ext-irita-wasm.gasLimit"), extConfig.get("vscode-ext-irita-wasm.mnemonic"), extConfig.get("vscode-ext-irita-wasm.algorithm"));
 }
